feat(Holidays): 여러 휴일을 가지는 객체 필드 추가
