+++ src/main/java/attendance/domain/Holidays.java
@@ -1,0 +1,35 @@
+package attendance.domain;
+
+import java.time.DayOfWeek;
+import java.time.LocalDate;
+import java.time.format.TextStyle;
+import java.util.List;
+import java.util.Locale;
+
+public class Holidays {
+
+    private final List<Holiday> holidays;
+
+    public Holidays(final List<Holiday> holidays) {
+        this.holidays = holidays;
+    }
+
+    public static void validateAttendanceDate(final LocalDate attendanceDate) {
+        DayOfWeek dayOfWeek = attendanceDate.getDayOfWeek();
+        if (dayOfWeek == DayOfWeek.SATURDAY || dayOfWeek == DayOfWeek.SUNDAY) {
+            throw new IllegalArgumentException(formatErrorMessage(attendanceDate));
+        }
+    }
+
+    public List<Holiday> getHolidays() {
+        return holidays;
+    }
+
+    private static String formatErrorMessage(final LocalDate date) {
+        return String.format("[ERROR] %d월 %d일 %s은 등교일이 아닙니다.",
+                date.getMonthValue(),
+                date.getDayOfMonth(),
+                date.getDayOfWeek().getDisplayName(TextStyle.FULL, Locale.KOREA)
+        );
+    }
+}
